refactor(pipeline/scrape): extract AccountExtractor module
Phase 5 — Commit 5 of 9 (pipeline-decoupling-master-2026-05-28).

Carves the account-record extraction concern out of
ScrapeAutoMapper.ts into a dedicated sub-cluster of three files
under the per-cluster max-lines:150 eff cap. Each file owns a
single responsibility:

- AccountExtractor/RawShapeTrace.ts (27 eff)
  safeStringify, truncatePreview, BODY_PREVIEW_CHARS, traceRawShape.
  Best-effort log-only helper — emits the top-level keys and a
  JSON preview when extractAccountRecords surfaces zero items.
  Default export of traceRawShape (single public symbol).

- AccountExtractor/ContainerClaim.ts (101 eff)
  Suffix-matching claim engine: indexRecordKeys,
  pickAccountObjectsFromKey, probeContainerForWk,
  applyClaimAttempt, assignContainersInRecord,
  extractAllContainers, flattenContainersForLog, plus the
  looksLikeAccountRecord shape predicate used by both
  pickAccountObjectsFromKey AND rootAccountArray. Holds the
  longest-match-wins logic that prevents double-counting a
  `bankAccounts` array against the `accounts` WK key.

- AccountExtractor/AccountExtractor.ts (67 eff)
  Composes the three extraction tiers (named containers via
  ContainerClaim, txn-signature BFS via LifoCrawl, root-array
  fallback via rootAccountArray) plus the validated-identifier
  layer (isUsableIdentifier, pickIdFromField,
  extractValidIdentifier, extractAccountIds).

ScrapeAutoMapper.ts goes from 462 → 295 eff LoC (-167). The
public re-export block at the bottom now lists `extractAccountIds`,
`extractAccountRecords`, `isUsableIdentifier`, and
`extractAllContainers` alongside the existing surface — no
consumer needs to change its import path.

No production behaviour change. All 66 ScrapeAutoMapper-related
tests still pass unchanged (test bodies untouched per the
FROZEN-TESTS rule). Husky 15-gate hook green.
