+++ src/Scrapers/Pipeline/Mediator/Scrape/ScrapeAutoMapper.ts
@@ -23,12 +23,10 @@ import {
   type ApiRecord,
   DEFAULT_CURRENCY,
   type ScalarFieldHit,
-  type UntypedValue,
 } from './AutoMapperFacade/AutoMapperTypes.js';
 import {
   findAllFieldValues,
   findFieldValue,
-  flattenObjectTree,
   isSearchableObject,
   matchField,
 } from './BfsFieldSearch/BfsFieldSearch.js';
@@ -208,421 +206,6 @@ function autoMapTransaction(raw: ApiRecord): ITransaction | false {
     status: TransactionStatuses.Completed,
     identifier: idVal,
   };
-}
-
-/** Preview length for the raw body trace dump. */
-const BODY_PREVIEW_CHARS = 4096;
-
-/**
- * Stringify a response body, returning a short failure marker on throw.
- * @param body - API body.
- * @returns Full JSON or '<unstringifiable>'.
- */
-function safeStringify(body: ApiRecord): string {
-  try {
-    return JSON.stringify(body);
-  } catch {
-    return '<unstringifiable>';
-  }
-}
-
-/**
- * Truncate a JSON string for trace dumps.
- * @param json - Full JSON string.
- * @returns Truncated preview.
- */
-function truncatePreview(json: string): string {
-  if (json.length <= BODY_PREVIEW_CHARS) return json;
-  return `${json.slice(0, BODY_PREVIEW_CHARS)}…`;
-}
-
-/**
- * Trace-dump the raw response shape when extraction fails. Helps diagnose
- * bank-specific API formats (e.g. Hapoalim) without stopping the pipeline.
- * @param responseBody - The raw API body that yielded zero items.
- * @returns Always true (side-effect only).
- */
-function traceRawShape(responseBody: ApiRecord): true {
-  const topLevelKeys = Object.keys(responseBody);
-  const json = safeStringify(responseBody);
-  const preview = truncatePreview(json);
-  LOG.trace({
-    message: 'extractAccountRecords: 0 items — raw body shape',
-    topLevelKeys,
-    preview,
-  });
-  return true;
-}
-
-/**
- * Check if a value is a plain object with at least one
- * `WK_ACCT.id` field (any of the queryId / displayId aliases).
- * Used to recognize account-shape records that don't carry
- * txn-signature fields (e.g. Hapoalim's /general/accounts:
- * `[{bankNumber,branchNumber,accountNumber,…}]`).
- * @param v - Candidate value.
- * @returns True when v looks like an account record.
- */
-function looksLikeAccountRecord(v: UntypedValue): boolean {
-  if (!isSearchableObject(v)) return false;
-  const hit = findFieldValue(v as ApiRecord, WK_ACCT.id);
-  return hit !== false;
-}
-
-/**
- * Root-array fallback: if the response body is already an array of
- * account-shaped records, return it directly. Covers responses like
- * Hapoalim's /general/accounts which is [{bankNumber,accountNumber,…}]
- * at root with no txn-signature fields to trip findFirstArray.
- * @param responseBody - Parsed JSON response body.
- * @returns Root array of account records, or empty.
- */
-function rootAccountArray(responseBody: ApiRecord): readonly ApiRecord[] {
-  if (!Array.isArray(responseBody)) return [];
-  const arr = responseBody as readonly unknown[];
-  if (arr.length === 0) return [];
-  if (!looksLikeAccountRecord(arr[0])) return [];
-  return arr.filter(looksLikeAccountRecord).map((v): ApiRecord => v as ApiRecord);
-}
-
-/**
- * Extracts every WK named container reachable from `responseBody` and
- * returns them keyed by container name. Phase 7d adds this helper so
- * ACCOUNT-RESOLVE.POST can commit BOTH `cards` AND `bankAccounts`
- * found in the same VisaCal `account/init` payload (the legacy
- * `findContainerArray` returned only the first match).
- *
- * <p>Container names that are absent from the body, or present but
- * empty after the {@link looksLikeAccountRecord} filter, are omitted
- * from the returned object entirely (so consumers can `Object.keys`
- * for a clean per-container count without zero-length entries).
- *
- * @param responseBody - Parsed JSON response body.
- * @returns Map of container-name → records. Empty when no WK
- *   container surfaced any account-shape records.
- */
-/** Per-record key index built once per record (preserves casing). */
-interface IRecordKeyIndex {
-  readonly originalKeys: readonly string[];
-  readonly lowerKeys: readonly string[];
-}
-
-/**
- * Build a per-record key index — pairs original-case keys with
- * lower-cased copies so each suffix probe avoids re-casing.
- * @param record - Object whose keys to index.
- * @returns Key index.
- */
-function indexRecordKeys(record: ApiRecord): IRecordKeyIndex {
-  const originalKeys = Object.keys(record);
-  const lowerKeys = originalKeys.map((k): string => k.toLowerCase());
-  return { originalKeys, lowerKeys };
-}
-
-/**
- * Filter the array under `record[originalKey]` to account-shape
- * records only. Returns the empty array when the value is not a
- * non-empty array of account-shape records — caller skips assigning.
- * @param record - Parent record.
- * @param originalKey - Key into the record.
- * @returns Account-shape records, possibly empty.
- */
-function pickAccountObjectsFromKey(record: ApiRecord, originalKey: string): readonly ApiRecord[] {
-  const value = record[originalKey];
-  if (!Array.isArray(value) || value.length === 0) return [];
-  const filtered = value.filter(looksLikeAccountRecord);
-  return filtered.map((v): ApiRecord => v as ApiRecord);
-}
-
-/** Per-WK probe outcome — successful claim or skip-with-reason. */
-interface IClaimAttempt {
-  readonly claimedLowerKey: string | false;
-  readonly objects: readonly ApiRecord[];
-}
-
-const NO_CLAIM: IClaimAttempt = { claimedLowerKey: false, objects: [] };
-
-/** Bundled args for {@link probeContainerForWk}. */
-interface IProbeArgs {
-  readonly record: ApiRecord;
-  readonly keys: IRecordKeyIndex;
-  readonly wantedLower: string;
-  readonly claimedLower: ReadonlySet<string>;
-}
-
-/**
- * Probe one record for the longest-suffix match of `wantedLower`
- * among unclaimed keys. Returns the claim metadata so the caller
- * mutates `assigned` and `claimedKeys` in one place — keeps the
- * loop body shallow per the project's max-depth=1 rule.
- * @param args - Bundled probe args (record, keys, wantedLower, claimedLower).
- * @returns Claim attempt metadata.
- */
-function probeContainerForWk(args: IProbeArgs): IClaimAttempt {
-  const matchIdx = args.keys.lowerKeys.findIndex(
-    (lk): boolean => !args.claimedLower.has(lk) && lk.endsWith(args.wantedLower),
-  );
-  if (matchIdx < 0) return NO_CLAIM;
-  const originalKey = args.keys.originalKeys[matchIdx];
-  const objects = pickAccountObjectsFromKey(args.record, originalKey);
-  if (objects.length === 0) return NO_CLAIM;
-  return { claimedLowerKey: args.keys.lowerKeys[matchIdx], objects };
-}
-
-/**
- * Walks `responseBody` once and assigns each PHYSICAL record-key
- * to the LONGEST WK name that suffix-matches it. Returns the
- * mutated `assigned` map so callers chain rather than rely on
- * implicit mutation. Without longest-match-wins, the suffix-matcher
- * in {@link tryContainerInRecord} would attribute the same
- * `bankAccounts` array to BOTH `accounts` (suffix hit) AND
- * `bankAccounts` (full hit), double-counting every record.
- *
- * @param record - One record from the body's flattened tree.
- * @param wkNames - WK container names sorted longest-first.
- * @param assigned - Per-WK-name records, mutated in place.
- * @returns The same `assigned` object (chain-friendly).
- */
-/** Bundled args for {@link applyClaimAttempt} so the call stays flat. */
-interface IApplyArgs {
-  readonly attempt: IClaimAttempt;
-  readonly wkName: string;
-  readonly assigned: Record<string, ApiRecord[]>;
-  readonly claimed: Set<string>;
-}
-
-/**
- * Apply one probe outcome — claim the lower-cased key and append
- * extracted records to the per-WK bucket. No-op when the attempt
- * surfaced no claim. Pulled out so the per-WK loop in
- * {@link assignContainersInRecord} stays flat (max-depth=1).
- * @param args - Bundled apply args.
- * @returns The same `assigned` map (chain-friendly).
- */
-function applyClaimAttempt(args: IApplyArgs): Record<string, ApiRecord[]> {
-  if (args.attempt.claimedLowerKey === false) return args.assigned;
-  args.claimed.add(args.attempt.claimedLowerKey);
-  const bucket = args.assigned[args.wkName] ?? [];
-  bucket.push(...args.attempt.objects);
-  args.assigned[args.wkName] = bucket;
-  return args.assigned;
-}
-
-/**
- * Walks `record` once and assigns each PHYSICAL key to the LONGEST
- * WK name that suffix-matches it. Without longest-match-wins, the
- * suffix-matcher would attribute the same `bankAccounts` array to
- * BOTH `accounts` (suffix hit) AND `bankAccounts` (full hit),
- * double-counting every record. Returns the mutated `assigned` map
- * so callers can chain.
- * @param record - One record from the body's flattened tree.
- * @param wkNames - WK container names sorted longest-first.
- * @param assigned - Per-WK-name records, mutated in place.
- * @returns The same `assigned` object (chain-friendly).
- */
-function assignContainersInRecord(
-  record: ApiRecord,
-  wkNames: readonly string[],
-  assigned: Record<string, ApiRecord[]>,
-): Record<string, ApiRecord[]> {
-  const keys = indexRecordKeys(record);
-  const claimed = new Set<string>();
-  for (const wkName of wkNames) {
-    const wantedLower = wkName.toLowerCase();
-    const attempt = probeContainerForWk({ record, keys, wantedLower, claimedLower: claimed });
-    applyClaimAttempt({ attempt, wkName, assigned, claimed });
-  }
-  return assigned;
-}
-
-/**
- * Extracts every WK named container reachable from `responseBody`
- * and returns them keyed by container name. Phase 7d adds this
- * helper so ACCOUNT-RESOLVE.POST can commit BOTH `cards` AND
- * `bankAccounts` found in the same VisaCal `account/init` payload
- * (the legacy `findContainerArray` returned only the first match).
- * @param responseBody - Parsed JSON response body.
- * @returns Per-WK-name container map; absent containers omitted.
- */
-function extractAllContainers(
-  responseBody: ApiRecord,
-): Readonly<Record<string, readonly ApiRecord[]>> {
-  const wkLongestFirst = [...WK_ACCT.containers].sort((a, b): number => b.length - a.length);
-  const assigned: Record<string, ApiRecord[]> = {};
-  const allRecords = flattenObjectTree(responseBody);
-  for (const r of allRecords) assignContainersInRecord(r, wkLongestFirst, assigned);
-  return assigned;
-}
-
-/**
- * Extract account records from API response. Logs the response shape
- * at trace level when zero items are found — exposes per-bank mapper
- * gaps. Tries three extractors in order:
- *
- * <ol>
- *   <li>{@link extractAllContainers} (named WK_ACCT.containers —
- *       cardsList, cards, accounts, bankAccounts; covers card-family
- *       banks where the cards array carries no txn-signature fields).
- *       Concatenates ALL surfaced containers — Phase 7d change.</li>
- *   <li>{@link findFirstArray} (txn-signature BFS — covers banks
- *       whose account response has txn preview arrays).</li>
- *   <li>{@link rootAccountArray} (root-level array of account-shaped
- *       records — covers Hapoalim's /general/accounts).</li>
- * </ol>
- *
- * @param responseBody - Parsed JSON response body.
- * @returns Account records with all original fields.
- */
-/**
- * Concatenate every container's records and emit the per-container
- * trace line. Extracted helper so {@link extractAccountRecords}
- * stays inside the project's cognitive-complexity ceiling.
- * @param containers - Per-WK-name container split.
- * @returns Concatenated records across every container.
- */
-function flattenContainersForLog(
-  containers: Readonly<Record<string, readonly ApiRecord[]>>,
-): readonly ApiRecord[] {
-  const containerNames = Object.keys(containers);
-  const concatenated: ApiRecord[] = [];
-  for (const name of containerNames) concatenated.push(...containers[name]);
-  LOG.debug({
-    message:
-      `extractAccountRecords: ${String(concatenated.length)} items ` +
-      `(${String(containerNames.length)} named containers: ${containerNames.join(',')})`,
-  });
-  return concatenated;
-}
-
-/**
- * Extract account records from API response. Logs the response
- * shape at trace level when zero items are found — exposes per-bank
- * mapper gaps. Tries three extractors in order:
- *
- * <ol>
- *   <li>{@link extractAllContainers} (named WK_ACCT.containers —
- *       cardsList, cards, accounts, bankAccounts; covers card-family
- *       banks where the cards array carries no txn-signature
- *       fields). Concatenates ALL surfaced containers — Phase 7d
- *       change.</li>
- *   <li>{@link findFirstArray} (txn-signature BFS — covers banks
- *       whose account response has txn preview arrays).</li>
- *   <li>{@link rootAccountArray} (root-level array of account-shaped
- *       records — covers Hapoalim's /general/accounts).</li>
- * </ol>
- *
- * @param responseBody - Parsed JSON response body.
- * @returns Account records with all original fields.
- */
-function extractAccountRecords(responseBody: ApiRecord): readonly ApiRecord[] {
-  const containers = extractAllContainers(responseBody);
-  if (Object.keys(containers).length > 0) return flattenContainersForLog(containers);
-  const items = findFirstArray(responseBody);
-  if (items.length > 0) {
-    LOG.debug({ message: `extractAccountRecords: ${String(items.length)} items` });
-    return castSearchable(items);
-  }
-  const rootAccts = rootAccountArray(responseBody);
-  if (rootAccts.length > 0) {
-    LOG.debug({
-      message: `extractAccountRecords: ${String(rootAccts.length)} items (root-array fallback)`,
-    });
-    return rootAccts;
-  }
-  LOG.debug({ message: 'extractAccountRecords: 0 items' });
-  traceRawShape(responseBody);
-  return [];
-}
-
-/**
- * Returns true when {@link id} is a real, server-accepted identifier
- * rather than a position index, sentinel, or stringification artifact.
- *
- * <p>The pipeline previously accepted any non-empty string as a valid
- * cardId, which let position indices (`cardIndex: 0` → `"0"`) and the
- * `'default'` credential sentinel slip through into transaction-API
- * URLs — the API silently returned 0 transactions and the pipeline
- * reported `success: true` with empty data. Per the Fail Fast
- * principle, identifiers must be validated at the extraction layer so
- * downstream stages either get a real ID or a loud failure.
- *
- * <p>Accepts: any string ≥ 2 chars that isn't a known sentinel
- * (`default`, `null`, `undefined`). The 2-char minimum rejects all
- * single-digit position values; no real banking identifier is < 2
- * characters in any captured trace across the 11 banks in the
- * registry.
- *
- * @param id - Candidate identifier string.
- * @returns True iff {@link id} is acceptable as a transaction-API
- *   query parameter.
- */
-function isUsableIdentifier(id: string): boolean {
-  if (id.length < 2) return false;
-  if (id === 'default') return false;
-  if (id === 'null') return false;
-  if (id === 'undefined') return false;
-  return true;
-}
-
-/**
- * Resolves a single account record to a usable identifier by walking
- * `WK_ACCT.id` (queryId fields first, displayId fields second) in
- * priority order and returning the first value that passes
- * {@link isUsableIdentifier}.
- *
- * <p>Lets banks whose response contains both a position-like field
- * (`cardIndex: 0`) AND a real identifier (`cardNumber: "0814"`)
- * deterministically yield the real one — important for Max-class
- * card-family banks where the SPA omits `cardUniqueId` from the
- * `getRegisterUserData` shape.
- *
- * @param record - One account/card record.
- * @returns First usable identifier, or empty string if none qualify.
- */
-/**
- * Resolves a single field name to a usable identifier within a record,
- * or returns the empty-string sentinel when the field is missing or
- * its value fails {@link isUsableIdentifier}. Extracted so
- * {@link extractValidIdentifier} stays inside the one-block max-depth
- * ceiling.
- *
- * @param record - One account/card record.
- * @param field - Single WK identifier field name to probe.
- * @returns Validated identifier or empty string.
- */
-function pickIdFromField(record: ApiRecord, field: string): string {
-  const value = findFieldValue(record, [field]);
-  if (value === false) return '';
-  const str = String(value);
-  if (!isUsableIdentifier(str)) return '';
-  return str;
-}
-
-/**
- * Resolves a single record to a usable identifier by walking
- * `WK_ACCT.id` (queryId fields, then displayId fields) and
- * returning the first value that passes {@link isUsableIdentifier}.
- * Drives the queryId-then-displayId fallback that lets card-family
- * banks like Max yield `cardNumber` when `cardUniqueId` is absent.
- *
- * @param record - One account/card record.
- * @returns First usable identifier, or empty-string sentinel.
- */
-function extractValidIdentifier(record: ApiRecord): string {
-  const candidates = WK_ACCT.id.map((field): string => pickIdFromField(record, field));
-  return candidates.find(Boolean) ?? '';
-}
-
-/**
- * Extracts validated account identifiers from an API response.
- *
- * @param responseBody - Parsed JSON response body.
- * @returns Array of usable identifier strings (empty when no record
- *   yielded a value passing {@link isUsableIdentifier}).
- */
-function extractAccountIds(responseBody: ApiRecord): readonly string[] {
-  const records = extractAccountRecords(responseBody);
-  return records.map(extractValidIdentifier).filter(Boolean);
 }
 
 /**
@@ -1037,16 +620,18 @@ export type {
 } from '../../Types/PipelineContext.js';
 export {
   autoMapTransaction,
-  extractAccountIds,
-  extractAccountRecords,
-  extractAllContainers,
   extractTransactions,
   extractTransactionsForCard,
   findAllFieldValues,
   findFieldValue,
   findFirstArray,
-  isUsableIdentifier,
   matchField,
   parseAutoDate,
   resolveTxnEndpoint,
 };
+export {
+  extractAccountIds,
+  extractAccountRecords,
+  extractAllContainers,
+  isUsableIdentifier,
+} from './AccountExtractor/AccountExtractor.js';
